(split)bug 97052  - use the same mechanism for turning on/off synchronous indexing in SOAP-based unit tests and in mock server unit tests  - manuall commits are used for synchronous index operation (imitating pre-9.0 behavior), each manual commit is a hard commit  - by default Solr commits will be done according to solrconfig.xml, not manually. Manual commits = synchronous indexing.  - manual commits are controlled by zimbra_index_manual_commit LC setting. Generally, this should only be turned on for testing. In production, admins should configure solrconfig.xml according to their needs  - add commit event listener Solr plugin that counts commits per Core. This plugin works in Embedded Solr as well as in Solr web app  - add a commit count handler to manipulate commit counter via Solr REST interface. This handler also works in Embedded Solr  - when manual commits are switched on, SolrIndex will wait for commit counter to go to zero before performing search  - remove Thread.sleep from unit tests where it was used to wait for Solr commits  - add a utility method to get an instance of SolrServer to TestUtil (used in new backup and restore SOAP tests)
TODO: manage indexing-related settings in solrconfig.xml by zmconfigd in non-AlwaysOn mode

https://bugzilla.zimbra.com/show_bug.cgi?id=97052

Copied from Perforce
 Change: 534002

+++ src/java/com/zimbra/qa/unittest/TestRedoLog.java
@@ -16,14 +16,21 @@
  */
 package com.zimbra.qa.unittest;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+
 import java.io.File;
 import java.io.PrintStream;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import junit.framework.TestCase;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
 
+import com.zimbra.common.localconfig.LC;
 import com.zimbra.common.util.ByteUtil;
 import com.zimbra.common.util.DevNullOutputStream;
 import com.zimbra.cs.account.Account;
@@ -39,18 +46,22 @@ import com.zimbra.cs.store.StoreManager;
 /**
  * Tests redolog operations
  */
-public class TestRedoLog
-extends TestCase {
-
-    private static final String USER_NAME = "user1";
-    private static final String RESTORED_NAME = "testredolog";
+public class TestRedoLog {
     private static final String NAME_PREFIX = TestRedoLog.class.getSimpleName();
+    private static final String USER_NAME = NAME_PREFIX + "_user1";
+    private static final String RESTORED_NAME = "testredolog";
+    private boolean originalLCSetting = false;
 
-    @Override public void setUp()
+    @Before
+    public void setUp()
     throws Exception {
         cleanUp();
+        originalLCSetting = LC.zimbra_index_manual_commit.booleanValue();
+        LC.zimbra_index_manual_commit.setDefault(true);
+        TestUtil.createAccount(USER_NAME);
     }
 
+    @Test
     public void testRedoLogVerify()
     throws Exception {
         RedoLogVerify verify = new RedoLogVerify(null, new PrintStream(new DevNullOutputStream()));
@@ -62,6 +73,7 @@ extends TestCase {
      * to another and leaves the original blob intact.  See bug 22873.
      */
 
+    @Test
     public void testTestRestoreMessageToNewAccount()
     throws Exception {
         // Add message to source account.
@@ -97,14 +109,20 @@ extends TestCase {
         return new File("/opt/zimbra/redolog/redo.log");
     }
 
-    @Override public void tearDown()
+    @After
+    public void tearDown()
     throws Exception {
         cleanUp();
+        LC.zimbra_index_manual_commit.setDefault(originalLCSetting);
     }
 
     private void cleanUp()
     throws Exception {
-        TestUtil.deleteTestData(USER_NAME, NAME_PREFIX);
-        TestUtil.deleteAccount(RESTORED_NAME);
+        if(TestUtil.accountExists(USER_NAME)) {
+            TestUtil.deleteAccount(USER_NAME);
+        }
+        if(TestUtil.accountExists(RESTORED_NAME)) {
+            TestUtil.deleteAccount(RESTORED_NAME);
+        }
     }
 }
